refactor: 테스트 클래스 네이밍 변경
